dto seggregation
