+++ src/main/java/com/thiru/investment_tracker/controller/CorporateActionController.java
@@ -1,6 +1,6 @@
 package com.thiru.investment_tracker.controller;
 
-import com.thiru.investment_tracker.dto.CorporateActionDto;
+import com.thiru.investment_tracker.dto.helper.CorporateActionDto;
 import com.thiru.investment_tracker.dto.user.UserMail;
 import com.thiru.investment_tracker.entity.CorporateActionEntity;
 import com.thiru.investment_tracker.service.CorporateActionService;
