+++ src/main/java/com/thiru/investment_tracker/controller/PortfolioController.java
@@ -3,9 +3,11 @@ package com.thiru.investment_tracker.controller;
 import com.thiru.investment_tracker.dto.*;
 import com.thiru.investment_tracker.dto.enums.HoldingType;
 import com.thiru.investment_tracker.dto.reports.profitloss.ProfitAndLossResponse;
+import com.thiru.investment_tracker.dto.request.AssetRequest;
+import com.thiru.investment_tracker.dto.request.BulkGetRequest;
+import com.thiru.investment_tracker.dto.response.AssetResponse;
 import com.thiru.investment_tracker.dto.user.UserMail;
 import com.thiru.investment_tracker.entity.AssetEntity;
-import com.thiru.investment_tracker.entity.ProfitAndLossEntity;
 import com.thiru.investment_tracker.entity.TransactionEntity;
 import com.thiru.investment_tracker.service.PortfolioService;
 import com.thiru.investment_tracker.service.TemporaryService;
