+++ src/main/java/com/thiru/investment_tracker/dto/context/AssetContext.java
@@ -1,4 +1,4 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.context;
 
 import java.time.LocalDate;
