+++ src/main/java/com/thiru/investment_tracker/dto/context/ReportContext.java
@@ -1,4 +1,4 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.context;
 
 import java.time.LocalDate;
