+++ src/main/java/com/thiru/investment_tracker/dto/helper/BrokerageChargesDto.java
@@ -1,16 +1,10 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.helper;
 
 
-import com.thiru.investment_tracker.dto.enums.BrokerName;
 import com.thiru.investment_tracker.dto.enums.BrokerageAggregatorType;
-import com.thiru.investment_tracker.dto.enums.EntityStatus;
-import com.thiru.investment_tracker.entity.model.BrokerageCharges;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.data.mongodb.core.mapping.Field;
-
-import java.time.LocalDate;
 
 @Data
 @AllArgsConstructor
