+++ src/main/java/com/thiru/investment_tracker/dto/helper/InputRecords.java
@@ -1,4 +1,4 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.helper;
 
 import java.util.ArrayList;
 import java.util.List;
