+++ src/main/java/com/thiru/investment_tracker/dto/model/TransactionEntityProtoType.java
@@ -1,4 +1,4 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.model;
 
 import com.thiru.investment_tracker.entity.TransactionEntity;
