+++ src/main/java/com/thiru/investment_tracker/dto/request/AssetManagementDetailsRequest.java
@@ -1,4 +1,4 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.request;
 
 import com.thiru.investment_tracker.dto.enums.AmcChargeFrequency;
 import com.thiru.investment_tracker.dto.enums.BrokerName;
