+++ src/main/java/com/thiru/investment_tracker/dto/request/BrokerChargesRequest.java
@@ -1,9 +1,9 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.request;
 
 
+import com.thiru.investment_tracker.dto.helper.BrokerageChargesDto;
 import com.thiru.investment_tracker.dto.enums.BrokerName;
 import com.thiru.investment_tracker.dto.enums.EntityStatus;
-import com.thiru.investment_tracker.entity.model.BrokerageCharges;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
