+++ src/main/java/com/thiru/investment_tracker/dto/request/BulkGetRequest.java
@@ -1,8 +1,9 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.request;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import com.thiru.investment_tracker.dto.helper.DateRange;
 import com.thiru.investment_tracker.entity.query.QueryFilter;
 
 import lombok.Data;
