+++ src/main/java/com/thiru/investment_tracker/dto/request/FinanceResponse.java
@@ -1,18 +1,18 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.request;
 
-import com.thiru.investment_tracker.dto.enums.CalculationType;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
+@Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class FinanceRequest {
-    private CalculationType calculationType;
+public class FinanceResponse {
     private double principalAmount;
     private double rateOfInterest;
     private int timePeriod;
     private double finalAmount;
-    private double emiAmount;
+    private int emiAmount;
 }
