+++ src/main/java/com/thiru/investment_tracker/dto/request/InsuranceRequest.java
@@ -1,36 +1,38 @@
-package com.thiru.investment_tracker.dto;
+package com.thiru.investment_tracker.dto.request;
 
 import java.util.List;
 
 import com.thiru.investment_tracker.dto.enums.InsuranceType;
 import com.thiru.investment_tracker.dto.enums.PolicyType;
-import com.thiru.investment_tracker.entity.model.PolicyDetails;
 
 import lombok.Data;
 
 @Data
-public class InsuranceResponse {
+public class InsuranceRequest {
 
 	private String email;
-	private String portFrom;
-	private String portTo;
 	private String insuranceId;
 	private String insurancePolicy;
 	private InsuranceType insuranceType;
 	private PolicyType policyType;
-	private double premium;
+
 	private String startDate;
 	private String endDate;
 	private String maturityDate;
 	private String renewalDate;
 	private String maturityAmount;
+
+	private double premium;
+
 	private String broker;
+	private String agent;
 	private String agentName;
 	private String agentEmail;
 	private String agentContact;
 	private String agentAddress;
+
 	private String insuranceStatus;
 	private String notes;
 
-	private List<PolicyDetails> policyDetails;
+	private List<String> persons;
 }
