+++ src/main/java/com/thiru/investment_tracker/entity/AssetEntity.java
@@ -2,7 +2,7 @@ package com.thiru.investment_tracker.entity;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.thiru.investment_tracker.dto.OrderTimeQuantity;
+import com.thiru.investment_tracker.dto.helper.OrderTimeQuantity;
 import com.thiru.investment_tracker.dto.enums.AccountType;
 import com.thiru.investment_tracker.dto.enums.AssetType;
 import com.thiru.investment_tracker.dto.enums.BrokerName;
