+++ src/main/java/com/thiru/investment_tracker/service/PortfolioService.java
@@ -1,18 +1,18 @@
 package com.thiru.investment_tracker.service;
 
-import com.thiru.investment_tracker.dto.AssetRequest;
-import com.thiru.investment_tracker.dto.AssetResponse;
-import com.thiru.investment_tracker.dto.OrderTimeQuantity;
-import com.thiru.investment_tracker.dto.ReportContext;
+import com.thiru.investment_tracker.dto.request.AssetRequest;
+import com.thiru.investment_tracker.dto.response.AssetResponse;
+import com.thiru.investment_tracker.dto.helper.OrderTimeQuantity;
+import com.thiru.investment_tracker.dto.context.ReportContext;
 import com.thiru.investment_tracker.dto.context.BuyContext;
 import com.thiru.investment_tracker.dto.enums.AccountType;
 import com.thiru.investment_tracker.dto.enums.AssetType;
 import com.thiru.investment_tracker.dto.enums.BrokerName;
 import com.thiru.investment_tracker.dto.enums.HoldingType;
 import com.thiru.investment_tracker.dto.enums.TransactionType;
-import com.thiru.investment_tracker.dto.reports.profitloss.ProfitAndLossContext;
+import com.thiru.investment_tracker.dto.context.ProfitAndLossContext;
 import com.thiru.investment_tracker.dto.reports.profitloss.ProfitAndLossResponse;
-import com.thiru.investment_tracker.dto.reports.profitloss.ProfitLossContext;
+import com.thiru.investment_tracker.dto.context.ProfitLossContext;
 import com.thiru.investment_tracker.dto.user.UserMail;
 import com.thiru.investment_tracker.entity.AssetEntity;
 import com.thiru.investment_tracker.entity.TemporaryTransactionEntity;
