+++ src/main/java/com/thiru/investment_tracker/service/ProfitAndLossService.java
@@ -8,9 +8,9 @@ import com.thiru.investment_tracker.dto.enums.AssetType;
 import com.thiru.investment_tracker.dto.enums.BrokerChargeTransactionType;
 import com.thiru.investment_tracker.dto.enums.CorporateActionType;
 import com.thiru.investment_tracker.dto.enums.TransactionType;
-import com.thiru.investment_tracker.dto.reports.profitloss.ProfitAndLossContext;
+import com.thiru.investment_tracker.dto.context.ProfitAndLossContext;
 import com.thiru.investment_tracker.dto.reports.profitloss.ProfitAndLossResponse;
-import com.thiru.investment_tracker.dto.reports.profitloss.ProfitLossContext;
+import com.thiru.investment_tracker.dto.context.ProfitLossContext;
 import com.thiru.investment_tracker.dto.user.UserMail;
 import com.thiru.investment_tracker.entity.ProfitAndLossEntity;
 import com.thiru.investment_tracker.entity.UserBrokerCharges;
