+++ src/main/java/com/thiru/investment_tracker/service/TransactionService.java
@@ -1,7 +1,7 @@
 package com.thiru.investment_tracker.service;
 
-import com.thiru.investment_tracker.dto.AssetRequest;
-import com.thiru.investment_tracker.dto.TransactionResponse;
+import com.thiru.investment_tracker.dto.request.AssetRequest;
+import com.thiru.investment_tracker.dto.response.TransactionResponse;
 import com.thiru.investment_tracker.dto.enums.AssetType;
 import com.thiru.investment_tracker.dto.enums.BrokerName;
 import com.thiru.investment_tracker.dto.user.UserMail;
