+++ src/main/java/com/thiru/investment_tracker/service/parser/model/RequestParser.java
@@ -1,6 +1,5 @@
 package com.thiru.investment_tracker.service.parser.model;
 
-import com.thiru.investment_tracker.dto.AssetRequest;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.util.List;
